refactor(#3405): split session_backend terminal-usage and stream-line clusters into child modules
#3344 pushed src/services/session_backend.rs to 1023 prod LoC and re-registered
it as a giant_file_registry entry (deadline 2026-08-31). This splits the two
clusters the issue named into sibling child modules under
src/services/session_backend/, following the #3038 S-series verbatim-move
discipline (logic byte-identical, only `use`/visibility paths adjusted), and
drops the root to 393 prod LoC — back under the 1000 threshold.

Child module structure (#[path] declarations + re-exports in the root preserve
every existing `crate::services::session_backend::X` path):

- session_backend/stream_line.rs (568 prod LoC) — the stream-line state machine
  cluster: `StreamLineState`/`TaskStartInfo`, `process_stream_line`, and the
  synchronous envelope parsers (`parse_stream_message`,
  `parse_stream_message_with_state`, `observe_stream_context`,
  `classify_task_notification_kind`, `emit_status_events_from_stream_json`,
  `parse_assistant_extra_tool_uses`, `json_str`, `claude_session_id`), plus the
  three `process_stream_line`-only tests
  (`assistant_text_plus_done_counts_text_only`,
  `housekeeping_only_window_harvests_nothing`,
  `process_stream_line_emits_workflow_status_events`).

- session_backend/terminal_usage.rs (106 prod LoC) — the #3344 terminal-usage
  adoption + analytics re-parser cluster: `MAX_PLAUSIBLE_PER_CALL_CONTEXT_TOKENS`,
  `adopt_terminal_result_usage` (shared by the watcher and the re-parser),
  `extract_turn_analytics_from_output{,_range}`, plus the three
  `extract_turn_analytics_*` analytics-path tests.

The root retains the process/IO + lifecycle layer: `ProcessBackend`/the
`SessionBackend` trait, the in-memory process-session registry, the
`read_output_file_until_result*` reader trio + `ReadHarvestStats`, and the four
reader-path tests. The two children reach each other
(`process_stream_line` ↔ `adopt_terminal_result_usage`,
`extract_turn_analytics_from_output_range` ↔ `process_stream_line`/`StreamLineState`)
through the root re-exports via their `use super::*`.

Verbatim verification: every moved prod function/struct/const compares
byte-identical (whitespace-normalized) against HEAD. The only moved-code edit is
a path adjustment in three relocated tests: `super::extract_turn_analytics_from_output_range`
→ unqualified, since the function now lives in the same `terminal_usage` module.
Original `pub(crate)` items are re-exported as `pub(crate) use`; `pub` items as
`pub use`. `TaskStartInfo`/`parse_stream_message` (no external consumer, the
latter has zero call sites) stay `pub` in the child but are not re-exported,
matching the S-series rule of not widening beyond what is consumed.

Registry/docs: removed the session_backend [[entry]] from
scripts/giant_file_registry.toml (it is no longer a prod-giant; keeping it would
trip the generator's ghost-registration check). Re-ran
scripts/generate_inventory_docs.py (ARCHITECTURE.md picks up the new child dir)
and moved the session_backend note in docs/agent-maintenance/change-surfaces.md
back into the decomposed-below-threshold list. session_backend was never in
audit_maintainability_giant_baseline.toml, so no ratchet edit was needed.

Gates (all green):
- cargo fmt --check
- cargo clippy --all-targets -- -D warnings
- cargo test --lib session_backend  (test result: ok. 10 passed)
- cargo test --lib tmux_output_stream  (test result: ok. 13 passed)
- cargo test --lib services::discord  (test result: ok. 1620 passed)
- python3 scripts/audit_maintainability.py --check  (giant_files 0, giant_file_ratchet 0)
- python3 scripts/check_agent_maintenance_docs.py  (freshness check passed)
