+++ ARCHITECTURE.md
@@ -714,6 +714,9 @@ src/
 │   │   ├── runtime_config.rs
 │   │   ├── session_control.rs
 │   │   └── store.rs
+│   ├── session_backend/
+│   │   ├── stream_line.rs
+│   │   └── terminal_usage.rs
 │   ├── slo/
 │   │   └── mod.rs
 │   ├── turn_orchestrator/
