+++ scripts/giant_file_registry.toml
@@ -327,15 +327,3 @@ file = "src/services/discord/turn_bridge/mod.rs"
 owner = "discord-relay"
 deadline = "2026-08-31"
 decompose_issue = "#3038"
-
-[[entry]]
-# Crossed the 1000-prod-LoC threshold (was 997, one line under) when #3344
-# round 2 landed the shared `adopt_terminal_result_usage` provenance/magnitude
-# gate plus its maintenance-rule doc — the gate is consumed by both the watcher
-# and the analytics re-parser, so it lives here next to `StreamLineState` /
-# `process_stream_line` rather than being duplicated. Active session-backend
-# surface; decompose the stream-parse + token-usage helpers out under #3036.
-file = "src/services/session_backend.rs"
-owner = "discord-relay"
-deadline = "2026-08-31"
-decompose_issue = "#3405"
